yt-WAWPP/WO/V2/work1/pyproject.toml
[project]
name = "work1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest-playwright>=0.5.2",
    "rich>=13.8.1",
]
